Update April-05-14 (UNTESTED)
The messages are now cleaned and sent using user messages.

Damage returning can now be turned off and is off by default. A warning
has been added due to the argument that happened over @ facepunch.

Added the ability to alert admins when a prop hits a player. (Please
note that the owner of the prop that is doing damage is not always a
prop killer.)

Changed the way variables are handled, and fixed a glitch with auto
blocking not turning off.